High score
Simple high score implementation that generates random scores in a range between 0 - 1000 and a random name from an array of names.

Save information will be loaded at the beginning of the first state of the game.

The project will log the chosen name and score in the Console regardless of whether it meets the criteria to save as a high score.

If the generated score exceeds the last saved score, it calls the method to save the new information.

Given the last state repeats itself, it will keep generating new names and scores in an attempt to beat the previous score.  I kept it for debugging purposes, but will be altering the state machine to include named transitions specific to where the game has progressed.

As before, the information is sent to basic GUILabels so that it can be seen without digging too deep into logs.  This will need to be removed before the project is finalized, or when it is no longer needed, as will the randomized content for the save info once game logic is implemented that will ask for a name at the beginning or end of the minigame.