scene-loading final push
Confirmation that all functions are working as designed so far to provide a stable framework for future programmers and designers on the project can work off of.

This branch utilizes a countdown timer to progress through a state machine that loads scenes as required based on specific criteria (most of which will have to be implemented once state logic starts being programmed out) and shows it in GUI Labels for easy accessibility.